python-services/insurance-edi/requirements.txt
# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2

# Async
asyncio==3.4.3
aiofiles==23.2.1

# Cryptography
pycryptodome==3.20.0
cryptography==41.0.7

# Network
httpx==0.26.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
